+++ src/extended-properties.spec.ts
@@ -4,8 +4,8 @@ import { Field } from './field';
 import { Group } from './group';
 import { List } from './list';
 import { transaction } from './transaction';
-import { Validator } from './validators';
 import { ValidationErrorText } from './validators/validation-error';
+import { Validator } from './validators/validator';
 
 /** what a UI layer attaches to an element: the properties it binds to the input it renders the element with */
 interface Presentation {