+++ src/validators/index.ts
@@ -1,4 +1,3 @@
 export * from './error-message-builder';
 export * from './validation-error';
-export * from './validator';
 export * as Validators from './index-validators';